[feature] ProductItem: polish product feature layout
